Add Bally squawk & talk boards (#5175)
* Bally pinball by35: Add sound LED to layouts.

* Bally pinball by35: whitespace/formatting cleanup.

* Bally pinball by35: Add filters to audio device output.

The boards have them, and the chip like to output DC at idle. This can
cause clipping in the mixer.

* 6821pia: For port A reads, only use the DDR for reading pins.

* Bally pinball by35: Move over cabinet switches.

This makes more room for a sound board LED.

* Bally pinball audio: Add Squawk & Talk board.

* Bally pinball audio: Add new board, Squawk & Talk with the AY chip.

* TMS5220: Adjust delay time between /RS going low and /READY being set.

The datasheet doesn't list this time exactly, but something close (time
between /RS going low and the data being valid on the bus). It's
measured in usec, not clock cycles.

* Bally audio: minor cleanup

* Midway mcr: Use Bally Squawk & Talk sound board implementation

* Bally pinball audio: remove clock from as2888, as it's a discrete board.