Pia6821 improvements (#5254)
* 6821pia: fix method names in comments

* zwackery: Remove PIA port A z_mask setting.

This would only be relevant if the PIA was read while in output mode.
These are inputs, and only read in input mode, so z_mask is not relevant.

* thomson: move pia pullup resistors from z_mask to read handler.

* 6821pia: remove unused set_port_a_z_mask calls.

* dgn_beta: move pia pullup resistors from z_mask to read handler.

* mpu3: Don't invert the VFD power on reset line.

It doesn't go through an inverter. I believe this was done to work
around a bug in the PIA port A implementation.

* pia6821: remove m_port_a_z_mask variable.

* 6821pia: For port A reads, only use the DDR for reading pins unless.

Provide an override option when a device depends on slightly undefined
behavior of an external device driving the pins and changing the value
read.

One board seems to need this (coinmstr). I don't have the schematics,
but it seems it uses this to check its meter operation.