Moving single stepping logic from mame_ui_manager to video_manager
The logic to step single frames was essentially implemented in
mame_ui_manager, and having this logic here has a negative impact on the
ability to use it in worker_ui, which provides an alternative
implementation of ui_manager.

This change also has the effect of removing the single step property
from luaengine, which is positive because in practice it should have
been seen as an implenentation detail.