mc6845 - configuration; row update paths; partial cursors.
1. Clarify the border offsets and make uses of the mc6845 driver consistent.

The most consistent definition of the border placement appears to be that the
displayable characters are always offset for the border and that only the
screen visual clip area changes with or without a border. With this definition
only the screen visual area changes when the machine chooses to display the
border or not, and the machines do not need special cases in either case. The
configuration is refactored into the mc6845 driver in most cases.

The other possible definition would have been to offset the displayable
characters from the origin of the screen pixmap only when displaying the
border, and otherwise place it at the origin. This would appear to place a
greater burden on each machine.

This change requires updates to the majority of the uses of the mc6845 driver
configuration paths and in particular the driver row update paths.

There are a large block of machines that do not use these paths. These
machines do not associate a screen with the mc6845 driver and do not use the
driver row update code paths, rather these machines implement their own screen
update function. This includes a lot of games. These are expected to be low
risk for regressions and only a selection have been tested.

There are a lot of machines that do use the driver configuration and row
update code paths, so this change affects a lot of machines. Most have been
tested except a few for which roms or software was not available or that were
just skeletons. For system that would at least configure the mc6845 these were
tested by writing to the display memory area.

In many cases it was only necessary to include a horizontal border offset in
the row update function. These are low risk for regression or at a least a low
risk for machine specific regressions.

Some machine row update functions included a vertical offset which was not
appropriate and was removed. The 'y' value passed to the row update function
is now clearly defined as an offset in the target bitmap so already includes
an offset for the border.

Many machines attempted to show the border. It is expected that this was
largely because the configuration was not working properly and the driver
definitions were frustrating. These have been modified to not show the border
which seems an appropriate default for these machines, and in any case these
should also work with the border shown. Showing the border, or not, might be
considered for a configuration option now as it simply changes the screen
visual clip area.

Many machines manually defined the screen size and visarea. In some cases they
did not even associate a screen with the mc6845 driver to avoid the driver
mis-configuring the screen. This is no longer necessary to most of them and
these definitions have been removed. The mc6845 driver should now correctly
configure the screen, and even if the machine runs some non-standard code
defining a different screen size etc. Most of these were tested and configured
properly and clipped the border properly.

A few machines had complex code to work around the configuration problems and
these have been clean up and can now rely on the mc6845 driver.

There remain only a couple of machines for which the driver configuration does
not work, and these appear to be targeting LCD screens etc that might not
follow CRT timing. These need to avoid associating a screen with the driver
and define their own screen size and visual area. Perhaps these would have
been better off using their own screen update functions rather than the driver
row update code paths, but these have been tested and still appear to work.

2. The mc6845 driver and a few machines mixed driver state with the
screen update path state, and this has been separated.

The screen update paths run off their the screen timers, and might update as
little as part of a row. So it is not appropriate for the driver or machine
state to be updating the driver state on screen updates. When the screen was
updating only part of a line, and calls the driver multiple times per line,
the driver was advancing an address on each causing artifacts - it was not
workable.

If some synchronisation is required between the screen updates and the machine
state then some more thought will be need to get this driver working with that
properly.

The mc6845 driver was updating the current display address from the screen
update code paths which appeared to be visible to the machine. The screen
update no longer modifies that machine state, calculating the display address
from the requested row.

The CGA driver stood out in this issue and need some rewriting. This raises
questions about the coordination of the screen update paths with the driver
state, the drivers row and sync state etc. Some graphical CGA code was test and
it appear to still run ok, but perhaps there had been some attempt to make the
screen updates more responsive for this driver, but it seemed only for the CGA
driver?

3. Partial cursors have been implement.

The mc6845 driver now implements partial cursors. It appears the only one
machine, the kaypro, had implemented partial cursors and that machines code
has been cleaned up now that the driver supports this.