SWTPC - add floppy disk support
Include the DC5 floppy disk controller on the expected SS-30 I/O positions of
ports 5 and 6.

Add a config option to select the CPU clock rate. Higher FDC clock rates need
higher CPU rates to download the data fast enough without DMA.

A couple of SWTBUG ROM patch config options are added to make disk
booting more practical.

Narrow the address range block used for the motherboard I/O to 0x8000 to
0x8fff, and also narrow the mirroring to potentially support two mother boards
with twice as many I/O cards. Add documentation to explain these changes.

Make 32K the default lower RAM size.

Implement RAM from 0xa000 to 0xdfff, as FLEX needs much of that.

Document the path to supporting Low and High PROMS.

Emit a 9600 baud rate clock on the 150/9600 line. This was a practical rate at
the time and screen based software is not practical at the prior rate of 1200
baud. This was a document option.

The emulator now runs FLEX 2 on a range of disk formats. Disk driver software
support is still an issue, another monitor PROM might be needed, there might
be an issue with the FDC 'ready' input, but this is a big step forward.