swtpc09: update, use the SS-30 bus, get these working.
Add support for the SS-30 bus. The DC5 FDC, MPS2, PIA-IDE and MPT have been
added to the SS-30 bus and are split out.

A separate floppy format has been added for UniFLEX, and the FLEX format has
been improved. These formats should be usable now, and have been lightly
tested.

The banked memory was not implemented correctly and has been updated.

The DMAF2 and DMAF3 FDCs are still implemented in the swtpc09 device, but have
been updated. These belong on the SS-50 bus, but there is no SS-50 bus yet -
still on the TODO list.

The PIA IDE has been implemented on the SS-30 bus, and updated to get it
working, and the swtpc09i machine is promoted to working. Have been able to
boot this from a FLEX hard disk image, and it also boots FLEX using the DC5
FDC.

Tested swtpc09 running FLEX with DMAF2 support. Tested swtpc09d3 running
UniFLEX and it seems to run well. The hard disk support is still TODO, but
added a reference to the WD1000 documentation. Did not have a UniFLEX image on
hand to test the swtpc09u, but have tested the DMAF2 under FLEX and UniFLEX
using DMAF3 so it's probably ok or close. Had no documentation for the DMAF3.