swtpc09: add hard disk support for the DMAF3 controller
This uses the WD1000 driver.

The DMAF2/3 6844 DMA support had to be improved to support multiple channels,
and now supports a separate high addresses for each channel. It now also
support addresses increasing or decreasing. These controllers use different
DMA channels for the floppy and the hard disk, and also a tape archiver which
is not yet implemented.

This gets UniFLEX running using a hard disk for storage, and does not appear
to regress the floppy support or other variants using the DMAF2.