[Imgtool] Fix handling multiple tokens in BASIC tokenizer
The line:
10 LIMIT$=STR$(LIMIT(N))
was not getting tokenised properly because the loop looking for tokens
wasn't exited and consumed consecutive tokens.

So $ was getting detected, token_shift and token_value were set, the
cursor position got incremented, then = got detected on the next
iteration of the loop.

We should instead exit the loop, and write what we already have.

Closes: #5478