MC6845: correct cursor.
The MC6845 shows a full cursor if the cursor end raster address is
greater than the maximum raster address, restore that case in general.

The issue on the BBC, that uses the HD6845S, might be the use of the
maximum raster address here, it appears to be programmed differently
on the HD6845S and also to depend on the interlaced mode. This patch
fixes the BBC while preserving the cursor emulation for the MC6845.