Remove Aliens: Neoplasma
This game is less than a year old, and the developers have released it
behind a, "How much would you like to donate?" barrier.  The tape
versions are playable in a browser as a demonstration, but to download
any tape or disk version, you must confirm the donation prompt.  The
TurboSound versions and disk versions are not even available as a demo
without passing the donation confirmation.

I'd rather people be encouraged to obtain this game from the official
sources.  Even if the conversion rate of demo plays to non-zero
donations is small, it's the principle of it.  The three-year rule
should be applied even-handedly.