fmtowns: bring machine configurations a bit closer to real hardware
- Separated the SCSI controller from the base machine configuration and
turned it into an optional device, since the first-generation models and
the Marty don't have one.

- Changed the CPU of the FM Towns II SJ to a 66 MHz 486DX2 instead of a
Pentium.

- Added comments describing the RAM expansion possibilities in each
model, and modified the user-selectable options accordingly.

- Added a separate machine configuration for the FM Towns II MX, since
it has a different CPU and more RAM options than the HR.

- Corrected memory maps for the 386SX-based models so the VRAM is
accessible at 0xb00000 in packed format. Fixes several instances of
graphics being garbled on the UX and Marty.

- Configured the Marty with just one floppy drive and 2 MB of RAM.

- Added a couple of sanity checks to prevent MAME from crashing when
selecting 1 MB of RAM or a single floppy drive. This makes it possible
to emulate something reasonably close to the functionality of an
original FM Towns Model 1, though the real Model 1/2 ROMs still need to
be dumped.