fmtowns: add the Model 2 ROMs
- Dumped and added the ROMs from an first-generation FM Towns Model 2,
and made it the parent set. For now we'll assume that the Model 1 ROMs
are the same, since the only differences between them are the amount of
floppy drives and RAM.

- Removed the old "fmtowns" set, since its ROMs completely match the FM
Towns II MX ones after splitting them.

- Renamed and added a few comments to the "fmtownsa" set so the
description is a bit more accurate. It seems to be from a
third-generation model, probaby a CX.

- Added a comment with a list of all the known FM Towns models.