ins8250: fix cleared pending THRE interrupt
The code notes that reading the iir register "will clear the int if
this is the source of the int" and the code cleared a pending THRE
interrupt if it was *pending*. This causes a lost THRE interrupt for
the SWTPC 8121 terminal firmware and this locks up the TX side of the
terminal. This occurred regularly when the ISR read this register upon
a higher priority RX pending interrupt. The documented operation has
been reinterpreted as only clearly the THRE interrupt if THRE is the
highest priority interrupt and so is the pending interrupt *causing*
the interrupt. This fixes that lost THRE interrupt for the SWTPC 8121
terminal.