IE15 terminal: rework the RS232 interfaces
The RS232 port has been moved out of the ie15 device. It is only needed in
stand alone operation, and not when connected to the rs232 bus, it was odd
leaving it in there, odd to still have an rs232 port slot option when running
stand alone.

The IE15 UART implementation has been removed from the rs232 bus code and back
into the ie15 device code. This leaves the rs232 bus code dealing with only
serial data, and not the parallel data, and that should make it easier in
future to have a common interface to RS232 terminals.