New working machines -------------------- RS-232 DCE-DCE Bridge
(nw) This is a simple machine for wiring together two things that want
to plug into an RS-232 port.  This means that for DCE-like things (e.g.
serial printers) there's no need to create an extra machine to allow
them to talk to the outside world with the null_modem device.  (Also
added some asserts, removed a FIXME, and put the swtpc8212 terminal in
the default RS-232 devices.)