new SOFTWARE LIST entry ----- gameking.xml : mc_4v19 - 4 in 1 - Warrior + Cleverhawk + Valiant + MetalDeform ('APR 23 05', 512KB cartridge)
the previous mc_4v19 has been made a clone mc_4v19a and description changed to "4 in 1 - Warrior + Cleverhawk + Valiant + MetalDeform ('JUN 16 05', 1MB cartridge)"

the new dump is playable, but from a different revision of the cartridge with only 512KB ROM instead of 1MB meaning the dump of this one is complete.  Presumably the 1MB versions exist either as bugfixes, or as a crude copy protecton against people with 512KB flash carts.