qvt102: key lines, boost interleave,
Key lines need to be logical 'anded' together, for the modifiers and the other
keys - if one is pulled low the line is pulled low.

Replace the use of set_perfect_quantum with a temporary boost_interleave to
meet the timing synchronisation between the keyboard MPU and the main CPU.

Move the resetting of the keyboard state from the reset method to the start
method. The keyboad MPU pushes values for these on reset anyway so it was
redundant.