Revert "8255: missed one driver (nw)"
This reverts commit f115aa8c665fe6b531cf344e0956ff23e7043def.

Revert "8255: set default tristate to 0 (nw)"

This reverts commit 716d06341dba3ab9ee883ce82abe1d4fe8202b8b.

This has been shown to break bw2, and it's very likely it breaks other
computers in subtle ways or in ways that are only noticeable when you
use a certain peripheral.  It's far too dangerous, particularly late in
a development cycle and with Christmas/new year coming.  You can't push
a change like this without checking every use of the device.

Please fix any drivers you subsequently modified to depend on the change.