new WORKING machines ---- Interactive TV Games 49-in-1 (PAL) [TeamEurope, David Haywood]
new NOT WORKING machines
----
Reactor MD (PAL) [TeamEurope, David Haywood]

the SunPlus part of the Reactor boots and runs, the MD part I haven't attempted to emulate, the main menu seems to run in an added / unsupported video mode when I took a brief look.  As such, it gets marked as not working, as the SunPlus part is only meant to boot once you select it from the MD based main menu.