Add partial dump of Kettle
I'm only adding dumps that completely passed tapclean's inspection,
after optimisation.  As this optimisation reduces wow and flutter
and other arbitrary timing aspects, it's reasonable to assume that
other people dumping the same tapes will be able to verify the present
sides and fill in the missing ones, which are oftentimes duplicates
anyway.