Don't clear the input queue on reset. (#6517)
This fixes problems where devices are input line changes are missed
because a device was reset. One case is at startup, where one device's
output is wire to another's input. During start the first device adds,
eg, a RESET assertion to the second. When that device starts, it resets
itself and clears the pending RESET assertion.