Allow embedded filesystems. [O. Galibert]
A file system is embedded by creating a C++ file through:
  scripts/build/makefs.py <name> file.cpp source-dir[:subdir-name] [source-dir[:..]...]

The contents in source-dir will be accessible in a virtual filesystem
at path subdir-name (default /).  The filesystem will be seen as an
opaque variable of type intfs::root_entry and name root_<name>.

Currently the only accessor is (in intfs.h):
  std::pair<std::unique_ptr<u8[]>, size_t> intfs::get_file(const root_entry &root, std::string filename);

which returns a pair with a unique_ptr to the data or nullptr if it is
not found and the size of the data.  The data is nul-terminated.

This is currently used, pretty much as a test, for layouts.  It has
the advantage of being a file generation and compilation like any
other, e.g. running in parallel with the others instead of
sequentially at the start of the compilation.

Possible TODO: find a way, if that's at all possible with
genie/makefiles, to automatically update the generated file if the
directory contents change.