z8000: Use callbacks to vector all external interrupts; correct data pushed during acknowledgment (nw)
For NMI and NVI, a vector is read from the bus but merely stored on the stack. (Most Z8000 systems seem to just leave the bus open here.) For VI, the vector offset calculation has been changed for the Z8001: the low byte is multiplied by 2, not 4, so AD0 generally must be driven low.