Created softlist for Timex cassettes (timex_cass)
Timex computer emulation is currently using the Spectrum softlist for cassettes, but many of those entries are actually incompatible unless the "Spectrum Emulator" cartridge is being used.
So I've created their own compatible (depending on the model) softlist, with every actual Timex dump I could find in World of Spectrum and Spectrum Computing.

These should work with the Timex "equivalents" to the Spectrum 48K and their clones:
-TS2068
-UK2086
-TC2068				(not currently emulated, or is it exactly the same as the UK2086?)
-Komputer 2086		(not currently emulated)

And some of them should work with the Timex equivalents of the Spectrum 16K:
-TC2048
-TC2048 (NTSC)		(not currently emulated)
-TS2048 (prototype)	(not currently preserved)

I've noticed that many of the TZX dumps were fake ones made from TAP files (so the real pauses between blocks were lost), so I just added the TAPs for those while any real TZXs are created.

Also, there were a few bad dumps of which I could only fix one at the moment, but they seem to load in Fuse anyway so I've left them in the softlist, labeled as such.

Also included some initial notes on the cassette dumps at the beginning of it, based on the spectrum_cass softlist but more to the point.

Somebody please hook this softlist to the systems in timex.cpp when possible. Thanks.