Namco custom chip improvements. (#6589)
* Wire up reset lines on namco custom chips.

* namco51: emulate via low level cpu

Previously, this was using high-level emulation. This had some
game-specific hacks.

* namco06: Emulate the NMI and chip select lines more accurately.

Three control bits are used for the clock divider.

* mb88xx: fix interrupt handling.

m_nf is basically the IF flag. It should be set when the int line is
active (when its low). Only trigger an int when the line was previously
high.

* namco custom chips: remove debug logging

* galaga, polepos: Remove quantum hacks.

With the better Namco custom chip emulation, this doesn't seem to be
required anymore.