Revert "ui: allow to skip warning popup at start"
This reverts commit 74f456df945cd1720dc4d6bc105f64c13e62ced8.

I will allow the feature with the following conditions:
* No skipping red warnings.
* A key-value map of device shortnames and yellow warnings must be
stored in the .cfg file, along with a timestamp of the last time the
system was run and the last time a warning was shown.
* On starting a system, the map of device shortnames and yellow warnings
must be built and compared to the one loaded from the .cfg file.  If it
doesn't match, if the system hasn't been started in seven days, or a
yellow warning hasn't been shown in 14 days, the warning is not to be
skipped.