mc6843: add a delay reading the CRC
Code using programmed I/O may read the last byte of a sector and then
expect to have some time to store that to memory before receiving an
interrupt on the completion of the command. The interrupt was
occurring on the last read and the last byte was being lost by some
drivers. This change adds a delay after reading the last byte, roughly
the time needed to read the 16 bit CRC, before the end of a read
command is processed.

Avoid side effects when these are disabled.