mdos dsk: support legacy floppy images, add XDOS support
Add legacy floppy image support which is currently required for the
mc6843 FDC emulator and used by the EXORset which run XDOS a variant
of MDOS.

Add support for 5.25 inch XDOS floppy drive formats as used by the
EXORset.

Make some of the identity tests more permissive to work with observer
disks. The date day and month are reversed on some disks, and the boot
rib cluster is lower.