Add MDCR support for P2000t
This adds support for the mini digital cassette recorder that can be
found inside a P2000t. This implementation is based on documentation
that can be found in https://github.com/p2000t/documentation.

I did some reverse engineering of the rom to discover:

- The RDC flip-flop is not well documented and just casually mentioned.
- There is no documentation that mentions the P2000 reads from tape when
the drive is moving in reverse.
- There is no mention that the RDC/RDA signals are swapped when
the tape is moving in reverse.

This emulation is not timing perfect, but within the bounds of what is
acceptable to the rom and diagnostics module.

In memory of NPM Jansen, who tought me all the magic of bits and bytes.