-Eliminate remaining <bezel> elements from internal layouts.
These layouts have changes to element stacking order that can't be
avoided without changing the group structure in the layout files.  I
think it's harmless, but it's possible it could have had a detrimental
effect on risc2500 (I tested it and didn't see any changes, but I might
not have known what to look for).

-complay.py: Added basic checks for dupicate collection names.

As with other checks of this kind, it doesn't actually instantiate the
layout so it doesn't check things when variable substitution is
involved.