Add MDCR support to P2000m
The P2000M shipped with an MDCR as well. We now turn the mdcr into
a required device and initialize it for the M model as well.

This fixes a nullptr exception upon access of ports 0x10, 0x20.

Includes a fix to reset the phase decoder when the end of tape signal is
raised.